project_0428/backend/requirements.txt
# 金融问答助手后端依赖
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
python-docx>=1.0.0
pdfplumber>=0.10.0
